simple tests working
